# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My First API",
        "version": "0.1"
    },
    "paths": {
...
        "/nilai-akademik/index": {
            "get": {
                "tags": [
                    "Nilai Akademik"
                ],
                "operationId": "96ec25bd29ff2fe98ed630380de01ccf",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/nilai-akademik/createorupdate": {
            "post": {
                "tags": [
                    "Nilai Akademik"
                ],
                "summary": "Create or update nilai Akademik",
                "description": "Create or update nilai Akademik.",
                "operationId": "478fbaf5d259f39205a8105b68041b9f",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "uuid": {
                                        "title": "uuid",
                                        "description": "UUID of nilai akademik (optional for update)",
                                        "type": "string",
                                        "example": ""
                                    },
                                    "name": {
                                        "title": "name",
                                        "description": "Name of nilai akademik",
                                        "type": "string",
                                        "example": "Aspek A"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/nilai-akademik/{id}/delete": {
            "delete": {
                "tags": [
                    "Nilai Akademik"
                ],
                "summary": "Delete nilai akademik",
                "description": "Delete nilai akademik by UUID.",
                "operationId": "840bb3564dd1f2123edc14fafbde74f8",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "UUID of nilai akademik to delete",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
...
    },
...
}